Separate uniter's Runner and Context into two packages. This will allow us to replace the context with other implementations.
