+++ worker/uniter/relationer.go
@@ -11,7 +11,7 @@ import (
 	apiuniter "github.com/juju/juju/api/uniter"
 	"github.com/juju/juju/worker/uniter/hook"
 	"github.com/juju/juju/worker/uniter/relation"
-	"github.com/juju/juju/worker/uniter/runner"
+	"github.com/juju/juju/worker/uniter/runner/context"
 )
 
 // Relationer manages a unit's presence in a relation.
@@ -34,13 +34,13 @@ func NewRelationer(ru *apiuniter.RelationUnit, dir *relation.StateDir, hooks cha
 }
 
 // ContextInfo returns a represention of r's current state.
-func (r *Relationer) ContextInfo() *runner.RelationInfo {
+func (r *Relationer) ContextInfo() *context.RelationInfo {
 	members := r.dir.State().Members
 	memberNames := make([]string, 0, len(members))
 	for memberName := range members {
 		memberNames = append(memberNames, memberName)
 	}
-	return &runner.RelationInfo{r.ru, memberNames}
+	return &context.RelationInfo{r.ru, memberNames}
 }
 
 // IsImplicit returns whether the local relation endpoint is implicit. Implicit
