+++ worker/uniter/runner/context/cache.go
@@ -1,7 +1,7 @@
 // Copyright 2012-2014 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package runner
+package context
 
 import (
 	"sort"
