+++ worker/uniter/runner/context/cache_test.go
@@ -1,7 +1,7 @@
 // Copyright 2012-2014 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package runner_test
+package context_test
 
 import (
 	"github.com/juju/errors"
@@ -10,7 +10,7 @@ import (
 	gc "gopkg.in/check.v1"
 
 	"github.com/juju/juju/apiserver/params"
-	"github.com/juju/juju/worker/uniter/runner"
+	"github.com/juju/juju/worker/uniter/runner/context"
 )
 
 type settingsResult struct {
@@ -38,19 +38,19 @@ func (s *RelationCacheSuite) ReadSettings(unitName string) (params.Settings, err
 }
 
 func (s *RelationCacheSuite) TestCreateEmpty(c *gc.C) {
-	cache := runner.NewRelationCache(s.ReadSettings, nil)
+	cache := context.NewRelationCache(s.ReadSettings, nil)
 	c.Assert(cache.MemberNames(), gc.HasLen, 0)
 	c.Assert(s.calls, gc.HasLen, 0)
 }
 
 func (s *RelationCacheSuite) TestCreateWithMembers(c *gc.C) {
-	cache := runner.NewRelationCache(s.ReadSettings, []string{"u/3", "u/2", "u/1"})
+	cache := context.NewRelationCache(s.ReadSettings, []string{"u/3", "u/2", "u/1"})
 	c.Assert(cache.MemberNames(), jc.DeepEquals, []string{"u/1", "u/2", "u/3"})
 	c.Assert(s.calls, gc.HasLen, 0)
 }
 
 func (s *RelationCacheSuite) TestInvalidateMemberChangesMembership(c *gc.C) {
-	cache := runner.NewRelationCache(s.ReadSettings, nil)
+	cache := context.NewRelationCache(s.ReadSettings, nil)
 	cache.InvalidateMember("foo/1")
 	c.Assert(cache.MemberNames(), jc.DeepEquals, []string{"foo/1"})
 	cache.InvalidateMember("foo/2")
@@ -61,7 +61,7 @@ func (s *RelationCacheSuite) TestInvalidateMemberChangesMembership(c *gc.C) {
 }
 
 func (s *RelationCacheSuite) TestRemoveMemberChangesMembership(c *gc.C) {
-	cache := runner.NewRelationCache(s.ReadSettings, []string{"x/2"})
+	cache := context.NewRelationCache(s.ReadSettings, []string{"x/2"})
 	cache.RemoveMember("x/1")
 	c.Assert(cache.MemberNames(), jc.DeepEquals, []string{"x/2"})
 	cache.RemoveMember("x/2")
@@ -70,7 +70,7 @@ func (s *RelationCacheSuite) TestRemoveMemberChangesMembership(c *gc.C) {
 }
 
 func (s *RelationCacheSuite) TestPruneChangesMembership(c *gc.C) {
-	cache := runner.NewRelationCache(s.ReadSettings, []string{"u/1", "u/2", "u/3"})
+	cache := context.NewRelationCache(s.ReadSettings, []string{"u/1", "u/2", "u/3"})
 	cache.Prune([]string{"u/3", "u/4", "u/5"})
 	c.Assert(cache.MemberNames(), jc.DeepEquals, []string{"u/3", "u/4", "u/5"})
 	c.Assert(s.calls, gc.HasLen, 0)
@@ -80,7 +80,7 @@ func (s *RelationCacheSuite) TestSettingsPropagatesError(c *gc.C) {
 	s.results = []settingsResult{{
 		nil, errors.New("blam"),
 	}}
-	cache := runner.NewRelationCache(s.ReadSettings, nil)
+	cache := context.NewRelationCache(s.ReadSettings, nil)
 
 	settings, err := cache.Settings("whatever")
 	c.Assert(settings, gc.IsNil)
@@ -92,7 +92,7 @@ func (s *RelationCacheSuite) TestSettingsCachesMemberSettings(c *gc.C) {
 	s.results = []settingsResult{{
 		params.Settings{"foo": "bar"}, nil,
 	}}
-	cache := runner.NewRelationCache(s.ReadSettings, []string{"x/2"})
+	cache := context.NewRelationCache(s.ReadSettings, []string{"x/2"})
 
 	for i := 0; i < 2; i++ {
 		settings, err := cache.Settings("x/2")
@@ -108,7 +108,7 @@ func (s *RelationCacheSuite) TestInvalidateMemberUncachesMemberSettings(c *gc.C)
 	}, {
 		params.Settings{"baz": "qux"}, nil,
 	}}
-	cache := runner.NewRelationCache(s.ReadSettings, []string{"x/2"})
+	cache := context.NewRelationCache(s.ReadSettings, []string{"x/2"})
 
 	settings, err := cache.Settings("x/2")
 	c.Assert(err, jc.ErrorIsNil)
@@ -128,7 +128,7 @@ func (s *RelationCacheSuite) TestInvalidateMemberUncachesOtherSettings(c *gc.C) 
 	}, {
 		params.Settings{"baz": "qux"}, nil,
 	}}
-	cache := runner.NewRelationCache(s.ReadSettings, nil)
+	cache := context.NewRelationCache(s.ReadSettings, nil)
 
 	settings, err := cache.Settings("x/2")
 	c.Assert(err, jc.ErrorIsNil)
@@ -148,7 +148,7 @@ func (s *RelationCacheSuite) TestRemoveMemberUncachesMemberSettings(c *gc.C) {
 	}, {
 		params.Settings{"baz": "qux"}, nil,
 	}}
-	cache := runner.NewRelationCache(s.ReadSettings, []string{"x/2"})
+	cache := context.NewRelationCache(s.ReadSettings, []string{"x/2"})
 
 	settings, err := cache.Settings("x/2")
 	c.Assert(err, jc.ErrorIsNil)
@@ -166,7 +166,7 @@ func (s *RelationCacheSuite) TestSettingsCachesOtherSettings(c *gc.C) {
 	s.results = []settingsResult{{
 		params.Settings{"foo": "bar"}, nil,
 	}}
-	cache := runner.NewRelationCache(s.ReadSettings, nil)
+	cache := context.NewRelationCache(s.ReadSettings, nil)
 
 	for i := 0; i < 2; i++ {
 		settings, err := cache.Settings("x/2")
@@ -180,7 +180,7 @@ func (s *RelationCacheSuite) TestPrunePreservesMemberSettings(c *gc.C) {
 	s.results = []settingsResult{{
 		params.Settings{"foo": "bar"}, nil,
 	}}
-	cache := runner.NewRelationCache(s.ReadSettings, []string{"foo/2"})
+	cache := context.NewRelationCache(s.ReadSettings, []string{"foo/2"})
 
 	settings, err := cache.Settings("foo/2")
 	c.Assert(err, jc.ErrorIsNil)
@@ -200,7 +200,7 @@ func (s *RelationCacheSuite) TestPruneUncachesOtherSettings(c *gc.C) {
 	}, {
 		params.Settings{"baz": "qux"}, nil,
 	}}
-	cache := runner.NewRelationCache(s.ReadSettings, nil)
+	cache := context.NewRelationCache(s.ReadSettings, nil)
 
 	settings, err := cache.Settings("x/2")
 	c.Assert(err, jc.ErrorIsNil)
