+++ worker/uniter/runner/errors.go
@@ -9,27 +9,7 @@ import (
 	"github.com/juju/errors"
 )
 
-var ErrRequeueAndReboot = errors.New("reboot now")
-var ErrReboot = errors.New("reboot after hook")
-var ErrNoProcess = errors.New("no process to kill")
 var ErrActionNotAvailable = errors.New("action no longer available")
-
-type missingHookError struct {
-	hookName string
-}
-
-func (e *missingHookError) Error() string {
-	return e.hookName + " does not exist"
-}
-
-func IsMissingHookError(err error) bool {
-	_, ok := err.(*missingHookError)
-	return ok
-}
-
-func NewMissingHookError(hookName string) error {
-	return &missingHookError{hookName}
-}
 
 type badActionError struct {
 	actionName string
