+++ worker/uniter/runner/factory_test.go
@@ -12,104 +12,35 @@ import (
 	"github.com/juju/names"
 	jc "github.com/juju/testing/checkers"
 	"github.com/juju/utils"
-	"github.com/juju/utils/fs"
 	gc "gopkg.in/check.v1"
 	"gopkg.in/juju/charm.v5/hooks"
 
-	"github.com/juju/juju/apiserver/params"
 	"github.com/juju/juju/state"
-	"github.com/juju/juju/storage"
-	"github.com/juju/juju/testcharms"
 	"github.com/juju/juju/worker/uniter/hook"
 	"github.com/juju/juju/worker/uniter/runner"
+	"github.com/juju/juju/worker/uniter/runner/context"
+	runnertesting "github.com/juju/juju/worker/uniter/runner/testing"
 )
 
 type FactorySuite struct {
-	HookContextSuite
-	paths      RealPaths
-	factory    runner.Factory
-	membership map[int][]string
+	ContextSuite
 }
 
 var _ = gc.Suite(&FactorySuite{})
-
-func (s *FactorySuite) SetUpTest(c *gc.C) {
-	s.HookContextSuite.SetUpTest(c)
-	s.paths = NewRealPaths(c)
-	s.membership = map[int][]string{}
-
-	contextFactory, err := runner.NewContextFactory(
-		s.uniter,
-		s.unit.Tag().(names.UnitTag),
-		fakeTracker{},
-		s.getRelationInfos,
-		s.storage,
-		s.paths,
-	)
-	c.Assert(err, jc.ErrorIsNil)
-
-	factory, err := runner.NewFactory(
-		s.uniter,
-		s.paths,
-		contextFactory,
-	)
-	c.Assert(err, jc.ErrorIsNil)
-	s.factory = factory
-}
-
-func (s *FactorySuite) SetCharm(c *gc.C, name string) {
-	err := os.RemoveAll(s.paths.charm)
-	c.Assert(err, jc.ErrorIsNil)
-	err = fs.Copy(testcharms.Repo.CharmDirPath(name), s.paths.charm)
-	c.Assert(err, jc.ErrorIsNil)
-}
-
-func (s *FactorySuite) getRelationInfos() map[int]*runner.RelationInfo {
-	info := map[int]*runner.RelationInfo{}
-	for relId, relUnit := range s.apiRelunits {
-		info[relId] = &runner.RelationInfo{
-			RelationUnit: relUnit,
-			MemberNames:  s.membership[relId],
-		}
-	}
-	return info
-}
-
-func (s *FactorySuite) setUpCacheMethods(c *gc.C) {
-	// The factory's caches are created lazily, so it doesn't have any at all to
-	// begin with. Creating and discarding a context lets us call updateCache
-	// without panicking. (IMO this is less invasive that making updateCache
-	// responsible for creating missing caches etc.)
-	_, err := s.factory.NewHookRunner(hook.Info{Kind: hooks.Install})
-	c.Assert(err, jc.ErrorIsNil)
-}
-
-func (s *FactorySuite) updateCache(relId int, unitName string, settings params.Settings) {
-	runner.UpdateCachedSettings(s.factory, relId, unitName, settings)
-}
-
-func (s *FactorySuite) getCache(relId int, unitName string) (params.Settings, bool) {
-	return runner.CachedSettings(s.factory, relId, unitName)
-}
 
 func (s *FactorySuite) AssertPaths(c *gc.C, rnr runner.Runner) {
 	c.Assert(runner.RunnerPaths(rnr), gc.DeepEquals, s.paths)
 }
 
 func (s *FactorySuite) TestNewCommandRunnerNoRelation(c *gc.C) {
-	rnr, err := s.factory.NewCommandRunner(runner.CommandInfo{RelationId: -1})
+	rnr, err := s.factory.NewCommandRunner(context.CommandInfo{RelationId: -1})
 	c.Assert(err, jc.ErrorIsNil)
 	s.AssertPaths(c, rnr)
-	ctx := rnr.Context()
-	s.AssertCoreContext(c, ctx)
-	s.AssertNotActionContext(c, ctx)
-	s.AssertNotRelationContext(c, ctx)
-	s.AssertNotStorageContext(c, ctx)
 }
 
 func (s *FactorySuite) TestNewCommandRunnerRelationIdDoesNotExist(c *gc.C) {
 	for _, value := range []bool{true, false} {
-		_, err := s.factory.NewCommandRunner(runner.CommandInfo{
+		_, err := s.factory.NewCommandRunner(context.CommandInfo{
 			RelationId: 12, ForceRemoteUnit: value,
 		})
 		c.Check(err, gc.ErrorMatches, `unknown relation id: 12`)
@@ -118,7 +49,7 @@ func (s *FactorySuite) TestNewCommandRunnerRelationIdDoesNotExist(c *gc.C) {
 
 func (s *FactorySuite) TestNewCommandRunnerRemoteUnitInvalid(c *gc.C) {
 	for _, value := range []bool{true, false} {
-		_, err := s.factory.NewCommandRunner(runner.CommandInfo{
+		_, err := s.factory.NewCommandRunner(context.CommandInfo{
 			RelationId: 0, RemoteUnitName: "blah", ForceRemoteUnit: value,
 		})
 		c.Check(err, gc.ErrorMatches, `invalid remote unit: blah`)
@@ -127,7 +58,7 @@ func (s *FactorySuite) TestNewCommandRunnerRemoteUnitInvalid(c *gc.C) {
 
 func (s *FactorySuite) TestNewCommandRunnerRemoteUnitInappropriate(c *gc.C) {
 	for _, value := range []bool{true, false} {
-		_, err := s.factory.NewCommandRunner(runner.CommandInfo{
+		_, err := s.factory.NewCommandRunner(context.CommandInfo{
 			RelationId: -1, RemoteUnitName: "blah/123", ForceRemoteUnit: value,
 		})
 		c.Check(err, gc.ErrorMatches, `remote unit provided without a relation: blah/123`)
@@ -135,70 +66,50 @@ func (s *FactorySuite) TestNewCommandRunnerRemoteUnitInappropriate(c *gc.C) {
 }
 
 func (s *FactorySuite) TestNewCommandRunnerEmptyRelation(c *gc.C) {
-	_, err := s.factory.NewCommandRunner(runner.CommandInfo{RelationId: 1})
+	_, err := s.factory.NewCommandRunner(context.CommandInfo{RelationId: 1})
 	c.Check(err, gc.ErrorMatches, `cannot infer remote unit in empty relation 1`)
 }
 
 func (s *FactorySuite) TestNewCommandRunnerRemoteUnitAmbiguous(c *gc.C) {
 	s.membership[1] = []string{"foo/0", "foo/1"}
-	_, err := s.factory.NewCommandRunner(runner.CommandInfo{RelationId: 1})
+	_, err := s.factory.NewCommandRunner(context.CommandInfo{RelationId: 1})
 	c.Check(err, gc.ErrorMatches, `ambiguous remote unit; possibilities are \[foo/0 foo/1\]`)
 }
 
 func (s *FactorySuite) TestNewCommandRunnerRemoteUnitMissing(c *gc.C) {
 	s.membership[0] = []string{"foo/0", "foo/1"}
-	_, err := s.factory.NewCommandRunner(runner.CommandInfo{
+	_, err := s.factory.NewCommandRunner(context.CommandInfo{
 		RelationId: 0, RemoteUnitName: "blah/123",
 	})
 	c.Check(err, gc.ErrorMatches, `unknown remote unit blah/123; possibilities are \[foo/0 foo/1\]`)
 }
 
 func (s *FactorySuite) TestNewCommandRunnerForceNoRemoteUnit(c *gc.C) {
-	rnr, err := s.factory.NewCommandRunner(runner.CommandInfo{
+	rnr, err := s.factory.NewCommandRunner(context.CommandInfo{
 		RelationId: 0, ForceRemoteUnit: true,
 	})
 	c.Assert(err, jc.ErrorIsNil)
 	s.AssertPaths(c, rnr)
-	ctx := rnr.Context()
-	s.AssertCoreContext(c, ctx)
-	s.AssertNotActionContext(c, ctx)
-	s.AssertRelationContext(c, ctx, 0, "")
-	s.AssertNotStorageContext(c, ctx)
 }
 
 func (s *FactorySuite) TestNewCommandRunnerForceRemoteUnitMissing(c *gc.C) {
-	rnr, err := s.factory.NewCommandRunner(runner.CommandInfo{
+	_, err := s.factory.NewCommandRunner(context.CommandInfo{
 		RelationId: 0, RemoteUnitName: "blah/123", ForceRemoteUnit: true,
 	})
 	c.Assert(err, gc.IsNil)
-	ctx := rnr.Context()
-	s.AssertCoreContext(c, ctx)
-	s.AssertNotActionContext(c, ctx)
-	s.AssertRelationContext(c, ctx, 0, "blah/123")
-	s.AssertNotStorageContext(c, ctx)
 }
 
 func (s *FactorySuite) TestNewCommandRunnerInferRemoteUnit(c *gc.C) {
 	s.membership[0] = []string{"foo/2"}
-	rnr, err := s.factory.NewCommandRunner(runner.CommandInfo{RelationId: 0})
+	rnr, err := s.factory.NewCommandRunner(context.CommandInfo{RelationId: 0})
 	c.Assert(err, jc.ErrorIsNil)
 	s.AssertPaths(c, rnr)
-	ctx := rnr.Context()
-	s.AssertCoreContext(c, ctx)
-	s.AssertNotActionContext(c, ctx)
-	s.AssertRelationContext(c, ctx, 0, "foo/2")
-	s.AssertNotStorageContext(c, ctx)
 }
 
 func (s *FactorySuite) TestNewHookRunner(c *gc.C) {
 	rnr, err := s.factory.NewHookRunner(hook.Info{Kind: hooks.ConfigChanged})
 	c.Assert(err, jc.ErrorIsNil)
 	s.AssertPaths(c, rnr)
-	ctx := rnr.Context()
-	s.AssertCoreContext(c, ctx)
-	s.AssertNotActionContext(c, ctx)
-	s.AssertNotRelationContext(c, ctx)
-	s.AssertNotStorageContext(c, ctx)
 }
 
 func (s *FactorySuite) TestNewHookRunnerWithBadHook(c *gc.C) {
@@ -248,10 +159,10 @@ func (s *FactorySuite) TestNewHookRunnerWithStorage(c *gc.C) {
 	uniter, err := st.Uniter()
 	c.Assert(err, jc.ErrorIsNil)
 
-	contextFactory, err := runner.NewContextFactory(
+	contextFactory, err := context.NewContextFactory(
 		uniter,
 		unit.Tag().(names.UnitTag),
-		fakeTracker{},
+		runnertesting.FakeTracker{},
 		s.getRelationInfos,
 		s.storage,
 		s.paths,
@@ -272,12 +183,6 @@ func (s *FactorySuite) TestNewHookRunnerWithStorage(c *gc.C) {
 	s.AssertPaths(c, rnr)
 	ctx := rnr.Context()
 	c.Assert(ctx.UnitName(), gc.Equals, "storage-block/0")
-	s.AssertStorageContext(c, ctx, "data/0", storage.StorageAttachmentInfo{
-		Kind:     storage.StorageKindBlock,
-		Location: "/dev/sdb",
-	})
-	s.AssertNotActionContext(c, ctx)
-	s.AssertNotRelationContext(c, ctx)
 }
 
 func (s *FactorySuite) TestNewHookRunnerWithRelation(c *gc.C) {
@@ -287,160 +192,6 @@ func (s *FactorySuite) TestNewHookRunnerWithRelation(c *gc.C) {
 	})
 	c.Assert(err, jc.ErrorIsNil)
 	s.AssertPaths(c, rnr)
-	ctx := rnr.Context()
-	s.AssertCoreContext(c, ctx)
-	s.AssertNotActionContext(c, ctx)
-	s.AssertRelationContext(c, ctx, 1, "")
-	s.AssertNotStorageContext(c, ctx)
-}
-
-func (s *FactorySuite) TestNewHookRunnerPrunesNonMemberCaches(c *gc.C) {
-
-	// Write cached member settings for a member and a non-member.
-	s.setUpCacheMethods(c)
-	s.membership[0] = []string{"rel0/0"}
-	s.updateCache(0, "rel0/0", params.Settings{"keep": "me"})
-	s.updateCache(0, "rel0/1", params.Settings{"drop": "me"})
-
-	rnr, err := s.factory.NewHookRunner(hook.Info{Kind: hooks.Install})
-	c.Assert(err, jc.ErrorIsNil)
-	s.AssertPaths(c, rnr)
-	ctx := rnr.Context()
-
-	settings0, found := s.getCache(0, "rel0/0")
-	c.Assert(found, jc.IsTrue)
-	c.Assert(settings0, jc.DeepEquals, params.Settings{"keep": "me"})
-
-	settings1, found := s.getCache(0, "rel0/1")
-	c.Assert(found, jc.IsFalse)
-	c.Assert(settings1, gc.IsNil)
-
-	// Check the caches are being used by the context relations.
-	relCtx, found := ctx.Relation(0)
-	c.Assert(found, jc.IsTrue)
-
-	// Verify that the settings really were cached by trying to look them up.
-	// Nothing's really in scope, so the call would fail if they weren't.
-	settings0, err = relCtx.ReadSettings("rel0/0")
-	c.Assert(err, jc.ErrorIsNil)
-	c.Assert(settings0, jc.DeepEquals, params.Settings{"keep": "me"})
-
-	// Verify that the non-member settings were purged by looking them up and
-	// checking for the expected error.
-	settings1, err = relCtx.ReadSettings("rel0/1")
-	c.Assert(settings1, gc.IsNil)
-	c.Assert(err, gc.ErrorMatches, "permission denied")
-}
-
-func (s *FactorySuite) TestNewHookRunnerRelationJoinedUpdatesRelationContextAndCaches(c *gc.C) {
-	// Write some cached settings for r/0, so we can verify the cache gets cleared.
-	s.setUpCacheMethods(c)
-	s.membership[1] = []string{"r/0"}
-	s.updateCache(1, "r/0", params.Settings{"foo": "bar"})
-
-	rnr, err := s.factory.NewHookRunner(hook.Info{
-		Kind:       hooks.RelationJoined,
-		RelationId: 1,
-		RemoteUnit: "r/0",
-	})
-	c.Assert(err, jc.ErrorIsNil)
-	s.AssertPaths(c, rnr)
-	ctx := rnr.Context()
-	s.AssertCoreContext(c, ctx)
-	s.AssertNotActionContext(c, ctx)
-	s.AssertNotStorageContext(c, ctx)
-	rel := s.AssertRelationContext(c, ctx, 1, "r/0")
-	c.Assert(rel.UnitNames(), jc.DeepEquals, []string{"r/0"})
-	cached0, member := s.getCache(1, "r/0")
-	c.Assert(cached0, gc.IsNil)
-	c.Assert(member, jc.IsTrue)
-}
-
-func (s *FactorySuite) TestNewHookRunnerRelationChangedUpdatesRelationContextAndCaches(c *gc.C) {
-	// Update member settings to have actual values, so we can check that
-	// the change for r/4 clears its cache but leaves r/0's alone.
-	s.setUpCacheMethods(c)
-	s.membership[1] = []string{"r/0", "r/4"}
-	s.updateCache(1, "r/0", params.Settings{"foo": "bar"})
-	s.updateCache(1, "r/4", params.Settings{"baz": "qux"})
-
-	rnr, err := s.factory.NewHookRunner(hook.Info{
-		Kind:       hooks.RelationChanged,
-		RelationId: 1,
-		RemoteUnit: "r/4",
-	})
-	c.Assert(err, jc.ErrorIsNil)
-	s.AssertPaths(c, rnr)
-	ctx := rnr.Context()
-	s.AssertCoreContext(c, ctx)
-	s.AssertNotActionContext(c, ctx)
-	s.AssertNotStorageContext(c, ctx)
-	rel := s.AssertRelationContext(c, ctx, 1, "r/4")
-	c.Assert(rel.UnitNames(), jc.DeepEquals, []string{"r/0", "r/4"})
-	cached0, member := s.getCache(1, "r/0")
-	c.Assert(cached0, jc.DeepEquals, params.Settings{"foo": "bar"})
-	c.Assert(member, jc.IsTrue)
-	cached4, member := s.getCache(1, "r/4")
-	c.Assert(cached4, gc.IsNil)
-	c.Assert(member, jc.IsTrue)
-}
-
-func (s *FactorySuite) TestNewHookRunnerRelationDepartedUpdatesRelationContextAndCaches(c *gc.C) {
-	// Update member settings to have actual values, so we can check that
-	// the depart for r/0 leaves r/4's cache alone (while discarding r/0's).
-	s.setUpCacheMethods(c)
-	s.membership[1] = []string{"r/0", "r/4"}
-	s.updateCache(1, "r/0", params.Settings{"foo": "bar"})
-	s.updateCache(1, "r/4", params.Settings{"baz": "qux"})
-
-	rnr, err := s.factory.NewHookRunner(hook.Info{
-		Kind:       hooks.RelationDeparted,
-		RelationId: 1,
-		RemoteUnit: "r/0",
-	})
-	c.Assert(err, jc.ErrorIsNil)
-	s.AssertPaths(c, rnr)
-	ctx := rnr.Context()
-	s.AssertCoreContext(c, ctx)
-	s.AssertNotActionContext(c, ctx)
-	s.AssertNotStorageContext(c, ctx)
-	rel := s.AssertRelationContext(c, ctx, 1, "r/0")
-	c.Assert(rel.UnitNames(), jc.DeepEquals, []string{"r/4"})
-	cached0, member := s.getCache(1, "r/0")
-	c.Assert(cached0, gc.IsNil)
-	c.Assert(member, jc.IsFalse)
-	cached4, member := s.getCache(1, "r/4")
-	c.Assert(cached4, jc.DeepEquals, params.Settings{"baz": "qux"})
-	c.Assert(member, jc.IsTrue)
-}
-
-func (s *FactorySuite) TestNewHookRunnerRelationBrokenRetainsCaches(c *gc.C) {
-	// Note that this is bizarre and unrealistic, because we would never usually
-	// run relation-broken on a non-empty relation. But verfying that the settings
-	// stick around allows us to verify that there's no special handling for that
-	// hook -- as there should not be, because the relation caches will be discarded
-	// for the *next* hook, which will be constructed with the current set of known
-	// relations and ignore everything else.
-	s.setUpCacheMethods(c)
-	s.membership[1] = []string{"r/0", "r/4"}
-	s.updateCache(1, "r/0", params.Settings{"foo": "bar"})
-	s.updateCache(1, "r/4", params.Settings{"baz": "qux"})
-
-	rnr, err := s.factory.NewHookRunner(hook.Info{
-		Kind:       hooks.RelationBroken,
-		RelationId: 1,
-	})
-	c.Assert(err, jc.ErrorIsNil)
-	s.AssertPaths(c, rnr)
-	ctx := rnr.Context()
-	rel := s.AssertRelationContext(c, ctx, 1, "")
-	c.Assert(rel.UnitNames(), jc.DeepEquals, []string{"r/0", "r/4"})
-	cached0, member := s.getCache(1, "r/0")
-	c.Assert(cached0, jc.DeepEquals, params.Settings{"foo": "bar"})
-	c.Assert(member, jc.IsTrue)
-	cached4, member := s.getCache(1, "r/4")
-	c.Assert(cached4, jc.DeepEquals, params.Settings{"baz": "qux"})
-	c.Assert(member, jc.IsTrue)
 }
 
 func (s *FactorySuite) TestNewHookRunnerWithBadRelation(c *gc.C) {
@@ -501,7 +252,7 @@ func (s *FactorySuite) TestNewActionRunnerGood(c *gc.C) {
 	ctx := rnr.Context()
 	data, err := ctx.ActionData()
 	c.Assert(err, jc.ErrorIsNil)
-	c.Assert(data, jc.DeepEquals, &runner.ActionData{
+	c.Assert(data, jc.DeepEquals, &context.ActionData{
 		Name: "snapshot",
 		Tag:  action.ActionTag(),
 		Params: map[string]interface{}{
