+++ worker/uniter/runner/util_unix_test.go
@@ -9,7 +9,6 @@ package runner_test
 var (
 	// Platform specific hook name used in runner_test.go
 	hookName = "something-happened"
-
 	// Platform specific script used in runner_test.go
 	echoPidScript = "echo $$ > pid"
 )
